refactor: move middleware api to dedicated file
